Move mode methods to utils
